Project_1/requirements.txt
actionlib==1.14.0
angles==1.9.13
attrs==22.2.0
base-local-planner==1.17.3
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cv-bridge==1.16.2
cycler==0.11.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
dynamixel-sdk==3.7.51
exceptiongroup==1.1.0
fonttools==4.38.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
idna==3.4
image-geometry==1.16.2
imageio==2.24.0
iniconfig==2.0.0
interactive-markers==1.12.0
joblib==1.2.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.4
laser-geometry==1.6.7
matplotlib==3.6.3
message-filters==1.15.15
networkx==3.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytz==2022.7.1
PyWavelets==1.4.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.28.2
resource-retriever==1.12.7
rosbag==1.15.15
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.15
roslaunch==1.15.15
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.15
rosmake==1.15.8
rosmaster==1.15.15
rosmsg==1.15.15
rosnode==1.15.15
rosparam==1.15.15
rospy==1.15.15
rosserial-python==0.9.2
rosservice==1.15.15
rostest==1.15.15
rostopic==1.15.15
rosunit==1.15.8
roswtf==1.15.15
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rtabmap-ros==0.20.22
rviz==1.14.19
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
tf2-sensor-msgs==0.7.6
threadpoolctl==3.1.0
tifffile==2022.10.10
tomli==2.0.1
topic-tools==1.15.15
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchsummary==1.5.1
torchvision==0.14.1+cu116
turtlebot3-example==1.2.5
turtlebot3-teleop==1.2.5
typing-extensions==4.4.0
urllib3==1.26.14
xacro==1.14.14
